requirements.txt
requests
playwright
eventlet
schedule
cf_clearance
biliup
xhs
qrcode
loguru
pyyaml
python-dotenv